tools/rmmod: move main logic into helpers
Move code in do_rmmod into two helpers: one to get the
dirname ($root/MODULE_DIRECTORY/$kernel) and another to do the main
logic.

The first helper, get_module_dirname(), is also always called to
figure the dirname path and to avoid providing NULL to the kmod_new()
API. This is because kmod_new() always defaults to MODULE_DIRECTORY when
the dirname path is empty, and won't work when MODULE_ALTERNATIVE_DIRECTORY
will be used.

This is in preparation to perform the same two steps again for
the new MODULE_ALTERNATIVE_DIRECTORY option.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>